reorganise all thoses files
