+++ src/App.jsx
@@ -1,10 +1,8 @@
 import React from 'react';
 import './App.css';
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
-import GlobalContaineur from './components/GlobalContaineur';
-import Accueil from './components/Accueil';
-// import { BrowserRouter as Router /* Switch, Route */ } from 'react-router-dom';
-// import SideMenu from './components/sidemenu/SideMenu';
+import GlobalContaineur from './components/navigation/GlobalContaineur';
+import Accueil from './components/navigation/Accueil';
 
 class App extends React.Component {
   constructor(props) {
