--- src/components/TopCountry.css
@@ -1,18 +1,0 @@
-.display-country {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-around;
-  margin-top: 10%;
-}
-
-.display-county-name {
-  width: 30%;
-}
-
-.display-county-operation {
-  width: 30%;
-}
-
-.display-county-price {
-  width: 30%;
-}
