+++ src/components/cards/Resume.jsx
@@ -1,24 +1,24 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import './globalContainer.css';
+import './cards.css';
 
 const tendances = [
   {
     id: 1,
+    nom: 'Botox',
+    valeur: '69,122',
+    tendance: '56%'
+  },
+  {
+    id: 2,
     nom: 'Rhinoplastie',
     valeur: '14,056',
     tendance: '56%'
   },
   {
-    id: 2,
-    nom: 'Lifting',
-    valeur: '14,056',
-    tendance: '56%'
-  },
-  {
     id: 3,
-    nom: 'Botox',
-    valeur: '14,056',
+    nom: 'Lifting',
+    valeur: '38,243',
     tendance: '56%'
   }
 ];
@@ -37,7 +37,12 @@ function Resume() {
       ))}
     </ul>
   );
-  return <div className="Resume">{allResume}</div>;
+  return (
+    <div className="Resume">
+      <h3>Les opérations en vogue</h3>
+      {allResume}
+    </div>
+  );
 }
 
 Resume.propTypes = {
