--- src/components/component.jsx
@@ -1,1 +1,0 @@
-import React from 'react';
