--- src/components/courbe.css
@@ -1,4 +1,0 @@
-img {
-  width: 100%;
-  height: auto;
-}
